# backend/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/products": {
      "get": {
        "description": "Returns a paginated list of products matching filter criteria (power, price, brand, category, specs).",
        "operationId": "ProductsController_getProducts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max 100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search keyword",
            "schema": {
              "example": "Hyundai",
              "type": "string"
            }
          },
          {
            "name": "brandId",
            "required": false,
            "in": "query",
            "description": "Filter by brand UUID",
            "schema": {
              "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filter by category UUID",
            "schema": {
              "example": "019fa8bc-8f4d-7000-b366-e691f45cfb90",
              "type": "string"
            }
          },
          {
            "name": "priceMin",
            "required": false,
            "in": "query",
            "description": "Minimum price in VND",
            "schema": {
              "example": 10000000,
              "type": "number"
            }
          },
          {
            "name": "priceMax",
            "required": false,
            "in": "query",
            "description": "Maximum price in VND",
            "schema": {
              "example": 500000000,
              "type": "number"
            }
          },
          {
            "name": "powerKvaMin",
            "required": false,
            "in": "query",
            "description": "Minimum power rating in kVA",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "powerKvaMax",
            "required": false,
            "in": "query",
            "description": "Maximum power rating in kVA",
            "schema": {
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "voltage",
            "required": false,
            "in": "query",
            "description": "Filter by voltage string",
            "schema": {
              "example": "230V",
              "type": "string"
            }
          },
          {
            "name": "phase",
            "required": false,
            "in": "query",
            "schema": {
              "example": "3phase",
              "type": "string",
              "enum": ["1phase", "3phase", "multi_phase"]
            }
          },
          {
            "name": "fuelType",
            "required": false,
            "in": "query",
            "schema": {
              "example": "diesel",
              "type": "string",
              "enum": ["diesel", "gasoline", "gas"]
            }
          },
          {
            "name": "canopyType",
            "required": false,
            "in": "query",
            "schema": {
              "example": "silent",
              "type": "string",
              "enum": [
                "silent",
                "super_silent",
                "open_frame",
                "closed_case",
                "tower",
                "rackmount"
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "newest",
              "example": "newest",
              "type": "string",
              "enum": ["newest", "priceAsc", "priceDesc"]
            }
          },
          {
            "name": "minPower",
            "required": false,
            "in": "query",
            "description": "Minimum power in kVA (alias for powerKvaMin)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPower",
            "required": false,
            "in": "query",
            "description": "Maximum power in kVA (alias for powerKvaMax)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "engineBrand",
            "required": false,
            "in": "query",
            "description": "Filter by engine brand",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alternatorBrand",
            "required": false,
            "in": "query",
            "description": "Filter by alternator brand",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Stock/lifecycle status",
            "schema": {
              "type": "string",
              "enum": ["active", "outOfStock", "all"]
            }
          },
          {
            "name": "isQuoteOnly",
            "required": false,
            "in": "query",
            "description": "Filter products marked for quote only",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ProductResponseDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Too Many Requests)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/too-many-requests",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Rate limit exceeded. Please try again later.",
                  "instance": "/api/example",
                  "invalidParams": [],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          }
        },
        "summary": "List products with hybrid faceted search and pagination",
        "tags": ["products"]
      },
      "post": {
        "description": "Creates a new product with technical specifications and images.",
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProductResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failure (Bad Request)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/bad-request",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": "Submitted data format is invalid",
                  "instance": "/api/example",
                  "invalidParams": [
                    {
                      "name": "email",
                      "reason": "Invalid email address format"
                    }
                  ],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          },
          "409": {
            "description": "Resource conflict (Conflict)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/conflict",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "Email address already exists",
                  "instance": "/api/example",
                  "invalidParams": [],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create product (Admin only)",
        "tags": ["products"]
      }
    },
    "/products/metadata": {
      "get": {
        "description": "Returns available filter ranges (price, power) and facet counts for brands, categories, fuel types, and phases.",
        "operationId": "ProductsController_getMetadata",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProductMetadataResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Too Many Requests)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/too-many-requests",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Rate limit exceeded. Please try again later.",
                  "instance": "/api/example",
                  "invalidParams": [],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          }
        },
        "summary": "Get faceted filter metadata",
        "tags": ["products"]
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Returns full product details by UUID or URL slug.",
        "operationId": "ProductsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProductResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found (Not Found)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "User profile not found",
                  "instance": "/api/example",
                  "invalidParams": [],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Too Many Requests)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/too-many-requests",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Rate limit exceeded. Please try again later.",
                  "instance": "/api/example",
                  "invalidParams": [],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          }
        },
        "summary": "Get product by ID or slug",
        "tags": ["products"]
      },
      "put": {
        "description": "Updates an existing product by UUID.",
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProductResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failure (Bad Request)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/bad-request",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": "Submitted data format is invalid",
                  "instance": "/api/example",
                  "invalidParams": [
                    {
                      "name": "email",
                      "reason": "Invalid email address format"
                    }
                  ],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found (Not Found)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "User profile not found",
                  "instance": "/api/example",
                  "invalidParams": [],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          },
          "409": {
            "description": "Resource conflict (Conflict)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/conflict",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "Email address already exists",
                  "instance": "/api/example",
                  "invalidParams": [],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product (Admin only)",
        "tags": ["products"]
      },
      "delete": {
        "description": "Soft deletes an existing product by UUID.",
        "operationId": "ProductsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found (Not Found)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Rfc9457ErrorResponseDto"
                },
                "example": {
                  "type": "http://localhost:3000/errors/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "User profile not found",
                  "instance": "/api/example",
                  "invalidParams": [],
                  "timestamp": "2026-07-25T02:45:00.000Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete product (Admin only)",
        "tags": ["products"]
      }
    },
...
  },
  "info": {
    "title": "Hyundai E-Commerce API",
    "description": "Hyundai E-Commerce Platform API for Commercial Vehicles, B2B Dealer Quoting, Warehouse Management, and Multi-channel Payment Processing.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token in the format: Bearer <token>"
      },
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "InvalidParamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Invalid field name",
            "example": "email"
          },
          "reason": {
            "type": "string",
            "description": "Detailed reason for the validation error",
            "example": "Invalid email address format"
          }
        },
        "required": ["name", "reason"]
      },
      "Rfc9457ErrorResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Standard HTTP error type URI",
            "example": "http://localhost:3000/errors/bad-request"
          },
          "title": {
            "type": "string",
            "description": "Standard HTTP error title",
            "example": "Bad Request"
          },
          "status": {
            "type": "number",
            "description": "HTTP status code",
            "example": 400
          },
          "detail": {
            "type": "string",
            "description": "Detailed error message",
            "example": "Submitted data format is invalid"
          },
          "instance": {
            "type": "string",
            "description": "API request path that triggered the error",
            "example": "/auth/register"
          },
          "invalidParams": {
            "description": "List of invalid parameters that failed validation",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidParamDto"
            }
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp when error occurred in ISO 8601 format",
            "example": "2026-07-25T02:45:00.000Z"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "timestamp"
        ]
      },
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["success"]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "Current page index (1-based)",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Number of records per page",
            "example": 20
          },
          "total": {
            "type": "number",
            "description": "Total number of matching records",
            "example": 100
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of calculated pages",
            "example": 5
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Indicates if there is a next page available",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "description": "Indicates if there is a previous page available",
            "example": false
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages",
          "hasNextPage",
          "hasPrevPage"
        ]
      },
      "PaginatedApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": ["success", "meta"]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f"
          },
          "nameVi": {
            "type": "string",
            "example": "Máy phát điện Diesel Hyundai DHY65KSE 60kVA 3 Pha"
          },
          "nameEn": {
            "type": ["string", "null"],
            "example": "Hyundai DHY65KSE 60kVA 3-Phase Diesel Generator"
          },
          "slug": {
            "type": "string",
            "example": "may-phat-dien-diesel-hyundai-dhy65kse"
          },
          "price": {
            "type": "string",
            "example": "245000000.00"
          },
          "isQuoteOnly": {
            "type": "boolean",
            "example": false,
            "description": "Whether the product requires quotation request (price <= 0)"
          },
          "descriptionVi": {
            "type": ["object", "null"],
            "example": null
          },
          "descriptionEn": {
            "type": ["object", "null"],
            "example": null
          },
          "shortDescriptionVi": {
            "type": ["string", "null"],
            "example": "Máy phát điện 60kVA vỏ chống ồn"
          },
          "shortDescriptionEn": {
            "type": ["string", "null"],
            "example": "60kVA diesel generator"
          },
          "images": {
            "example": [
              "https://res.cloudinary.com/hyundai/image/upload/dhy65kse.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "brandId": {
            "type": ["string", "null"],
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f"
          },
          "categoryId": {
            "type": ["string", "null"],
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb90"
          },
          "productType": {
            "type": "string",
            "enum": ["generator", "ups", "ats", "accessory"],
            "example": "generator"
          },
          "powerKva": {
            "type": ["string", "null"],
            "example": "60.00"
          },
          "powerKw": {
            "type": ["string", "null"],
            "example": "48.00"
          },
          "standbyPowerKva": {
            "type": ["string", "null"],
            "example": "66.00"
          },
          "standbyPowerKw": {
            "type": ["string", "null"],
            "example": "52.80"
          },
          "phase": {
            "type": ["string", "null"],
            "enum": ["1phase", "3phase", "multi_phase", null],
            "example": "3phase"
          },
          "voltage": {
            "type": ["string", "null"],
            "example": "230/400V"
          },
          "frequency": {
            "type": ["number", "null"],
            "example": 50
          },
          "fuelType": {
            "type": ["string", "null"],
            "enum": ["diesel", "gasoline", "gas", null],
            "example": "diesel"
          },
          "canopyType": {
            "type": ["string", "null"],
            "enum": [
              "silent",
              "super_silent",
              "open_frame",
              "closed_case",
              "tower",
              "rackmount",
              null
            ],
            "example": "silent"
          },
          "startMethod": {
            "type": ["string", "null"],
            "enum": ["electric", "recoil", "remote", "auto_ats", null],
            "example": "electric"
          },
          "engineBrand": {
            "type": ["string", "null"],
            "example": "Hyundai"
          },
          "alternatorBrand": {
            "type": ["string", "null"],
            "example": "Hyundai"
          },
          "upsTopology": {
            "type": ["string", "null"],
            "enum": [
              "offline",
              "line_interactive",
              "online_double_conversion",
              null
            ]
          },
          "upsBatteryType": {
            "type": ["string", "null"],
            "enum": ["internal", "external", null]
          },
          "specSheet": {
            "type": "object",
            "example": []
          },
          "specs": {
            "type": "object",
            "example": {
              "model": "DHY65KSE"
            }
          },
          "totalStockCache": {
            "type": "number",
            "example": 5
          },
          "totalSalesCache": {
            "type": "number",
            "example": 0
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "brand": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandResponseDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryResponseDto"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "nameVi",
          "slug",
          "price",
          "isQuoteOnly",
          "images",
          "specSheet",
          "specs",
          "totalStockCache",
          "totalSalesCache",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}